backend/requirements.txt
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
